Squash two more warnings from old compilers   Change an isprint to isgraph

+++ ntpd/nts_client.c
@@ -428,7 +428,7 @@ bool check_aead(SSL *ssl, struct peer* peer, const char *hostname) {
 		memcpy(buff, data, l);
 		buff[l] = '\0';
 		for (i=0; i<l; i++) {
-			if (!isgraph(buff[i])) {
+			if (!isgraph((int)buff[i])) {
 			  buff[i] = '*'; /* fix non-printing crap */
 			}
 		}
